requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
requests>=2.31.0
werkzeug==2.0.3
